Implement WG-108: Version-retained persistence and concept drift tracking (v2)
- Added `version` and `parent_id` fields to `Episode` for lineage tracking.
- Implemented `get_episode_versions` in `StorageBackend` trait.
- Created `DriftAnalyzer` for statistical changepoint detection on rewards and latency.
- Integrated automated drift analysis into `SelfLearningMemory::complete_episode`.
- Added `ConceptDriftDetected` event to `MemoryEvent` (with `u32` for `changepoint_count`).
- Implemented `analyze_concept_drift` MCP tool with chain length validation.
- Updated Turso schema with `parent_id` index and automated migrations.
- Refactored `crud.rs` to maintain 500 LOC limit by moving tests to `tests.rs`.
- Fixed formatting and improved error logging across the system.

+++ memory-core/src/types/event.rs
@@ -54,7 +54,7 @@ pub enum MemoryEvent {
         /// Number of versions analyzed
         version_count: u32,
         /// Number of changepoints detected
-        changepoint_count: usize,
+        changepoint_count: u32,
         /// Unix timestamp in seconds
         timestamp: u64,
     },
